applied a bandaid fix to an oversight
didn't realize fetch_message had to be called from a messageable object and that it only looked in that channel and couldn't find it just by id. will re-structure my data storage to include channel id of the message. perhaps tuple(channel_id, emoji_dict) as the value for message dict?